fix: fix private/approve page compilation error
